[WIP][VL] Support celeborn sort based shuffle
